ウィンドウの大きさを調整

+++ myproject/vec2.py
@@ -1,5 +1,5 @@
 class Vec2:
-    def __init__(self, x: float = 0, y: float = 0):
+    def __init__(self, x: int = 0, y: int = 0):
         self.x = x
         self.y = y
